requirements.txt
python-telegram-bot[job-queue]==20.0
requests>=2.28.1
cachetools>=5.2.0
humanize>=4.4.0
asana>=3.0.0
boto3>=1.26.44